Use the most common web API naming for a EventTarget's value, 'result.'
'result' seems to dominate most of the web apis. The idea that any of them might be considerable as a Future is too alluring to pass up. I reject all notions that this would be incorrect or potentially confusing: it should be, because it is, because there shouldn't be a difference.